% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{$,ACD-method}
\alias{$,ACD-method}
\title{Extract parts of ACD.}
\usage{
\S4method{$}{ACD}(x, name)
}
\arguments{
\item{x}{an object of ACD class}

\item{name}{member function of ACD class}
}
\description{
Extract parts of ACD.
}

