% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{$,HPC-method}
\alias{$,HPC-method}
\title{Extract parts of HPC.}
\usage{
\S4method{$}{HPC}(x, name)
}
\arguments{
\item{x}{an object of HPC class}

\item{name}{member function of HPC class}
}
\description{
Extract parts of HPC.
}

