% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{$,MCD-method}
\alias{$,MCD-method}
\title{Extract parts of MCD.}
\usage{
\S4method{$}{MCD}(x, name)
}
\arguments{
\item{x}{an object of MCD class}

\item{name}{member function of MCD class}
}
\description{
Extract parts of MCD.
}

