% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmcmControl.R
\name{jmcmControl}
\alias{jmcmControl}
\title{Control of Joint Mean Covariance Model Fitting}
\usage{
jmcmControl(trace = FALSE, profile = TRUE, errorMsg = FALSE)
}
\arguments{
\item{trace}{whether or not the value of the objective function and the
parameters should be print on every trace'th iteration.}

\item{profile}{whether or not parameters should be estimated sequentially
using the idea of profile likelihood.}

\item{errorMsg}{whether or not the error message should be print}
}
\description{
Construct control structures for joint mean covariance model
fitting
}

