% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{jmcmMod-class}
\alias{jmcmMod-class}
\title{Class "jmcmMod" of Fitted Joint Mean-Covariance Models.}
\description{
Class "jmcmMod" of Fitted Joint Mean-Covariance Models.
}
\section{Slots}{

\describe{
\item{\code{call}}{the matched call}

\item{\code{opt}}{the optimization result returned by optimizeJmcm}

\item{\code{args}}{arguments m, Y, X, Z, W, time}

\item{\code{triple}}{an integer vector of length three containing the degrees of the
three polynomial functions for the mean structure, the log innovation
-variances and the autoregressive or moving average coefficients when
'mcd' or 'acd' is specified for cov.method. It refers to the mean structure,
variances and angles when 'hpc' is specified for cov.method.}

\item{\code{devcomp}}{the deviance components list}
}}

