% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{meanplot}
\alias{meanplot}
\title{Plot Fitted Mean Curves}
\usage{
meanplot(object)
}
\arguments{
\item{object}{a fitted joint mean covariance model of class "jmcmMod", i.e.,
typically the result of jmcm().}
}
\description{
plot fitted mean curves
}
\examples{
cattleA <- cattle[cattle$group=='A', ]
fit.mcd <- jmcm(weight | id | I(ceiling(day/14 + 1)) ~ 1 | 1, data=cattleA,
  triple = c(8, 3, 4), cov.method = 'mcd')
meanplot(fit.mcd)

}
