% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmcm.R
\docType{methods}
\name{show,jmcmMod-method}
\alias{show,jmcmMod-method}
\title{Print information for jmcmMod-class}
\usage{
\S4method{show}{jmcmMod}(object)
}
\arguments{
\item{object}{a fitted joint mean covariance model of class "jmcmMod", i.e.,
typically the result of jmcm().}
}
\description{
Print information for jmcmMod-class
}
