% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmcmControl.R
\name{jmcmControl}
\alias{jmcmControl}
\title{Control of Joint Mean Covariance Model Fitting}
\usage{
jmcmControl(trace = FALSE, profile = TRUE, ignore.const.term = TRUE,
  original.poly.order = FALSE, errormsg = FALSE)
}
\arguments{
\item{trace}{whether or not the value of the objective function and the
parameters should be print on every trace'th iteration.}

\item{profile}{whether or not parameters should be estimated sequentially
using the idea of profile likelihood.}

\item{ignore.const.term}{whether or not the constant term should be considered 
when calculating log-likelihood and BIC.
function}

\item{original.poly.order}{whether or not the original poly order p q d should 
be used}

\item{errormsg}{whether or not the error message should be print}
}
\description{
Construct control structures for joint mean covariance model
fitting
}
